+++ jsprimer.net/todoapp/src/App.js
@@ -1,6 +1,6 @@
 import { TodoListModel } from "./model/TodoListModel.js";
+import { TodoListView } from "./view/TodoListView.js";
 import { TodoItemModel } from "./model/TodoItemModel.js";
-import { TodoItemView } from "./view/TodoItemView.js";
 import { element, render } from "./view/html-util.js";
 
 export class App {
@@ -17,10 +17,11 @@ export class App {
 
     // 2. TodoListModelの状態が更新されたら表示を更新する
     this.todoListModel.onChange(() => {
+      const todoListView = new TodoListView();
       // それぞれのTodoItem要素をtodoListElement以下へ追加する
       const todoItems = this.todoListModel.getTodoItems();
       // TodoリストをまとめるList要素
-      const todoListElement = this.createListElement(todoItems, {
+      const todoListElement = todoListView.createElement(todoItems, {
           // 指定したTodoアイテムの完了状態を反転させる
           onUpdateTodo: ({id, completed}) => {
             this.todoListModel.updateTodo({
@@ -53,28 +54,5 @@ export class App {
       inputElement.value = "";
     });
   }
-
-
-  /**
-   * `todoItems`に対応するTodoリストのHTML要素を作成して返す
-   * @param {TodoItemModel[]} todoItems TodoItemModelの配列
-   * @param {function({id:string, completed: boolean})} onUpdateTodo チェックボックスの更新イベントリスナー
-   * @param {function({id:string})} onDeleteTodo 削除ボタンのクリックイベントリスナー
-   * @returns {Element} TodoItemModelの配列に対応したリストのHTML要素
-   */
-  createListElement(todoItems, { onUpdateTodo, onDeleteTodo }) {
-    const todoListElement = element`<ul />`;
-    // 各TodoItemモデルに対応したHTML要素を作成し、リスト要素へ追加する
-    todoItems.forEach(todoItem => {
-      const todoItemView = new TodoItemView();
-      const todoItemElement = todoItemView.createElement(todoItem, {
-        onDeleteTodo,
-        onUpdateTodo
-      });
-      todoListElement.appendChild(todoItemElement);
-    });
-    return todoListElement;
-  }
-
 
 }
